split subiquity.core into subiquity.client.client and subiquity.server.server
